+++ src/model/CustomerGroup.java
@@ -7,15 +7,15 @@ package model;
 public class CustomerGroup {
 	private String name;
 	private String type;
-	private int maxDiscount;
+	private double maxDiscount;
 	
-	public CustomerGroup(String name, String type, int maxDiscount) {
+	public CustomerGroup(String name, String type, double maxDiscount) {
 		this.name = name;
 		this.type = type;
 		this.maxDiscount = maxDiscount;
 	}
 
-	public int getMaxDiscount() {
+	public double getMaxDiscount() {
 		return maxDiscount;
 	}
