+++ src/model/ShelfProduct.java
@@ -6,7 +6,7 @@ package model;
 public class ShelfProduct extends AbstractProduct implements SellableIF {
 	private AbstractProduct abstractProduct;
 	
-	public ShelfProduct(String name, String descripton, String barcode, double purchasePrice, double price, int discount) {
+	public ShelfProduct(String name, String descripton, String barcode, double purchasePrice, double price, double discount) {
 		super(name, descripton, barcode, purchasePrice, price, discount);
 	}
